requirements.txt
backcall==0.1.0
backports.csv==1.0.7
certifi==2019.6.16
chardet==3.0.4
decorator==4.4.0
defusedxml==0.6.0
dialite==0.5.2
diff-match-patch==20181111
Django==2.2.5
django-allauth==0.39.1
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap3==11.0.0
django-bootstrap4==0.0.8
django-braces==1.13.0
django-extensions==2.1.9
django-filter==2.1.0
django-import-export==1.2.0
django-oauth-toolkit==1.2.0
django-pdb==0.6.2
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
django-tables2-bootstrap4==2.0.0b5
djangorestframework==3.9.4
docopt==0.6.2
entrypoints==0.3
et-xmlfile==1.0.1
flake8==3.7.7
flexx==0.8.0
idna==2.8
ipython==7.6.1
ipython-genutils==0.2.0
isort==4.3.20
jdcal==1.4.1
jedi==0.14.0
mccabe==0.6.1
numpy==1.16.4
oauthlib==3.0.1
odfpy==1.4.0
openpyxl==2.6.2
pandas==0.24.2
parso==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
pipreqs==0.4.9
pkg-resources==0.0.0
prompt-toolkit==1.0.16
pscript==0.7.1
ptipython==1.0.1
ptpdb==0.17
ptpython==0.33
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.4.2
PyJWT==1.7.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.2
PyYAML==5.1.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
tablib==0.13.0
tornado==6.0.3
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
webruntime==0.5.6
xlrd==1.2.0
xlwt==1.3.0
yapf==0.27.0
yarg==0.1.9
